camptensor-engine/app/requirements.txt
Django==3.1.5
gunicorn==20.0.4
psycopg2-binary==2.8.5
requests==2.25.1
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-rest-swagger==2.2.0
django-redis==4.12.1
django-filter==2.4.0
matplotlib==3.3.4
nltk==3.5
numpy==1.20.1
Pillow==8.1.0
scipy==1.6.1
eventlet==0.30.2
celery==4.4.7
flower==0.9.7
gevent==21.1.2